+++ NEWS.rst
@@ -13,6 +13,7 @@ Changed
 - Make all completion function return codes accurate,
   complying with the shell's internal expectations
 - Don't append lines which are already present (pipa, pipac, pipacs)
+- Use zsh/mapfile to avoid six more subshells
 
 0.3.6
 =====